pyproject.toml
[project]
name = "llmembeddinganalysis-3rd"
version = "0.1.0"
description = "Add your description here"
dependencies = [
    "pip>=24.1.2",
    "accelerate==0.31.0",
    "datasets==2.21.0",
    "diskcache==5.6.3",
    "giotto-tda==0.6.2",
    "homcloud==4.4.1",
    "matplotlib==3.9.0",
    "numpy==1.26.1",
    "pandas==2.2.2",
    "plotly==5.22.0",
    "polars==0.20.31",
    "scikit-learn==1.3.2",
    "scipy==1.14.0",
    "seaborn==0.13.2",
    "sentence-transformers==3.0.1",
    "toml==0.10.2",
    "torch==2.4.0",
    "tqdm==4.66.5",
    "umap==0.1.1",
    "hydra-core>=1.3.2",
    "optuna>=3.6.1",
    "mysqlclient>=2.2.4",
    "pipdeptree>=2.23.1",
    "jupyterlab>=4.2.4",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/llmembeddinganalysis_3rd"]
